+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Ability to provide a custom matcher function to computed property assertions
 - Ability for computed property assertions to assert truthyness when no
   arguments are provided
+- Displaying computed selectors in assertion errors
 
 ### Fixed